pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zinkwell"
version = "0.1.0"
description = "Cross-platform Bluetooth photo printer library for Zink-based printers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Zinkwell Contributors"}
]
keywords = ["bluetooth", "printer", "zink", "photo", "canon", "ivy", "kodak"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: System :: Hardware",
]
dependencies = [
    "Pillow>=9.0.0",
    "loguru>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/jbgh2/bt-printer"
Repository = "https://github.com/jbgh2/bt-printer"

[tool.setuptools.packages.find]
include = ["zinkwell*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "hardware: tests requiring physical printer (skip in CI)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["zinkwell"]
branch = true
omit = [
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "if __name__ == .__main__.:",
]
show_missing = true
skip_covered = false
fail_under = 0

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
